docs: improve CentOS 7 and Ubuntu 16 rrdcached installation instructions (#7473)
I tried to get these to as few of steps as possible.
Ubuntu starts the service during installation, so it won't have proper permissions/settings until after you restart it.
CentOS doesn't need epel (it doesn't have rrdtool in it) or permissions fixes.  Move run files to /run directory to save a lot of steps and be similar to other platforms. Skipping the touch allows us to skip daemon-reload