Doc debian 10 updates (#11488)
* add package acl to required list

There are ways to install debian without the acl package. Add it to the
required list

* replace systemctl with more generic service

Debian does not require systemd. This makes calling systemctl to restart
processes a problem. The more generic way in Debian is to call service

* rm command allows multiple parameters

* use nginx config in a more Debian way

In Debian webservers are configured using -available and -enabled
directories. Make better use of this stucture.

* better snmpd.conf handling

Instead of overwriting snmpd.conf inform Debian of the change.

Because snmpd.conf contains sensitive info make it only user accessible

* replace nano with users preferred editor

* revert back to nano as editor

* remove dpkg-divert as per PR discussion